Finally being able to upload this project to git
I had such an issue trying to upload this project. I usually create the git repo before I even write a line of code but in this instance I thought I was not going to share this project at all. It was such an issue to upload a project that was already created but I found a work around that works for me.